+++ src/i18n/locales/es.ts
@@ -609,6 +609,7 @@ export const es = {
         ce_input_ph_id: 'DNI / Documento',
         ce_input_ph_address: 'Calle, ciudad, provincia',
         ce_social_which: '¿Qué red social es?',
+        ce_phone_apps: 'Mensajería:',
         ce_input_ph_other: 'Notas…',
         ce_input_ph_locality: 'Ciudad / Localidad',
         ce_alias_ph: 'Otro nombre por el que se conoce a esta persona',